Move blobstore type S3 test into separate file
